refactor(storage-manager): split `aligner.rs` file
The `aligner.rs` file was only growing in size, rendering navigation
complicated. This commit splits it in the following manner:
- `aligner_query.rs` inherits the `AlignmentQuery` and the logic that
  generates a reply to it.
- `aligner_reply.rs` inherits the `AlignmentReply` and the logic to
  process it.
- `core.rs` inherits the `spawn_query_replica_aligner` that spawns a
  task to query the "Aligner" of a Replica.

* plugins/zenoh-plugin-storage-manager/src/replication/aligner.rs:
  deleted and split the content of the file.

* plugins/zenoh-plugin-storage-manager/src/replication/core.rs:
  - Declared the new modules `aligner_query` and `aligner_reply`.
  - Updated the imports as the method `spawn_query_replica_aligner`
    was moved there.

* plugins/zenoh-plugin-storage-manager/src/replication/core/aligner_query.rs:
  new file that holds the logic to generate a reply to an AlignmentQuery.

* plugins/zenoh-plugin-storage-manager/src/replication/core/aligner_reply.rs:
  new file that holds the logic to process an AlignmentReply.

* plugins/zenoh-plugin-storage-manager/src/replication/mod.rs: removed
  the, no longer used, module `aligner`.
