+++ plugins/zenoh-plugin-storage-manager/src/replication/mod.rs
@@ -26,7 +26,6 @@
 //!
 //! [History::Latest]: zenoh_backend_traits::History::Latest
 
-mod aligner;
 mod classification;
 mod configuration;
 mod core;
